Move init functions to example_project, add pyspark install
